# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Закрытый API для системы управления перевозками",
        "description": "API documentation for My Application",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "andrew.izmaylov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://transportation.test/public/api/v1/",
            "description": "Local development server"
        },
        {
            "url": "https://transportation.test/public/api/v1/",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/shipper/{transportation_id}/add-transportation-address": {
            "post": {
                "tags": [
                    "3. Shipper - Адрес"
                ],
                "summary": "Добавление адреса",
                "description": "Регистрирует адрес забора/доставки для существующей перевозки. Заполняется отправителем груза. ID перевозки отправляется в URI",
                "operationId": "faf07485249cdfd8f16dd4dd2a6f2358",
                "parameters": [
                    {
                        "name": "transportation_id",
                        "in": "path",
                        "description": "Transportation ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для создания объекта адрес",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "alias",
                                    "type",
                                    "contact",
                                    "city",
                                    "addressLine1",
                                    "phoneNumber"
                                ],
                                "properties": {
                                    "alias": {
                                        "type": "string",
                                        "example": "Чемодан"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "delivery",
                                            "pickup"
                                        ],
                                        "example": "delivery"
                                    },
                                    "contact": {
                                        "type": "string",
                                        "example": "Иванов Петр Александрович"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Москва"
                                    },
                                    "addressLine1": {
                                        "type": "string",
                                        "example": "улица Строителей дом 5"
                                    },
                                    "addressLine2": {
                                        "type": "string",
                                        "example": "квартира 34"
                                    },
                                    "addressLine3": {
                                        "type": "string",
                                        "example": "домофон 15"
                                    },
                                    "phoneNumber": {
                                        "type": "string",
                                        "example": "+79789658965"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "RU"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Необходимые пояснения"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "TransportationAddress created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransportationCard"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/shipper/register-transportation": {
            "post": {
                "tags": [
                    "2. Shipper - Перевозки"
                ],
                "summary": "Создание заявки на перевозку",
                "description": "Регистрирует заявку на перевозку от лица зарегистрированного пользователя",
                "operationId": "3600f9cc301fe37b42db94c129c681f9",
                "requestBody": {
                    "description": "Данные для создания объекта перевозки",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "pickupFrom",
                                    "pickupTo"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Москва - Дубай, лекарства"
                                    },
                                    "pickupFrom": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-01-14 12:00:00"
                                    },
                                    "pickupTo": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-02-14 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transportation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransportationCard"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "TransportationCard": {
                "required": [
                    "id",
                    "type",
                    "attributes"
                ],
                "properties": {
                    "id": {
                        "description": "Transportation ID",
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "example": "Transportation"
                    },
                    "attributes": {
                        "required": [
                            "id",
                            "name",
                            "clientId",
                            "pickupFrom",
                            "pickupTo",
                            "transportationStatus"
                        ],
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "clientId": {
                                "type": "string"
                            },
                            "pickupFrom": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-01 10:00:00"
                            },
                            "pickupTo": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-01-01 18:00:00"
                            },
                            "transportationStatus": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}